Moved backglass manager into a separate package
