+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/backglassmanager/BackglassManagerDragDropHandler.java
@@ -1,4 +1,4 @@
-package de.mephisto.vpin.ui.tables.dialogs;
+package de.mephisto.vpin.ui.backglassmanager;
 
 import java.io.File;
 import java.util.Arrays;
@@ -13,11 +13,11 @@ import javafx.scene.layout.StackPane;
 
 public class BackglassManagerDragDropHandler extends BaseDragDropHandler {
   
-  BackglassManagerDialogController controller;
+  BackglassManagerController controller;
 
   private final List<String> suffixes = Arrays.asList( "zip", "rar", "directb2s" );
 
-  public BackglassManagerDragDropHandler(BackglassManagerDialogController controller, TableView<?> directb2sList, StackPane tableStack) {
+  public BackglassManagerDragDropHandler(BackglassManagerController controller, TableView<?> directb2sList, StackPane tableStack) {
     super(directb2sList, tableStack, true);
     this.controller = controller;
     this.overlayController.setMessage("Drop Backglass here...");
