+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/backglassmanager/BackglassManagerPredicateFactory.java
@@ -1,7 +1,7 @@
-package de.mephisto.vpin.ui.tables.dialogs;
+package de.mephisto.vpin.ui.backglassmanager;
 
 import de.mephisto.vpin.restclient.directb2s.DirectB2S;
-import de.mephisto.vpin.ui.tables.dialogs.BackglassManagerDialogController.DirectB2SEntryModel;
+import de.mephisto.vpin.ui.backglassmanager.BackglassManagerController.DirectB2SEntryModel;
 import de.mephisto.vpin.ui.tables.models.B2SVisibility;
 
 import org.apache.commons.lang3.StringUtils;
@@ -14,11 +14,11 @@ public class BackglassManagerPredicateFactory {
 
   private String filterValue;
 
-  private BackglassManagerDialogController controller;
+  private BackglassManagerController controller;
 
   private List<Integer> emulatorIds = new ArrayList<>();
 
-  public BackglassManagerPredicateFactory(BackglassManagerDialogController controller) {
+  public BackglassManagerPredicateFactory(BackglassManagerController controller) {
     this.controller = controller;
   }
